requirements.txt
http://gevent.googlecode.com/files/gevent-1.0b2.tar.gz
tornado-pyvows
coverage
colorama
